Fix compiler warnings across codebase
- Suppress -Wmissing-field-initializers and -Wmissing-braces in Makefile
- Include asset.h in sprite.h for complete Agentite_AssetHandle type
- Remove unused static functions in collision, noise, physics, msdf, ui files
- Add (void) casts for unused callback parameters in physics2d, gizmos, scene
- Remove unused variables in ui_style, ui_draw, collision
- Fix tautological comparison in state.cpp by casting enum to int

Build now compiles cleanly with -Wall -Wextra.

+++ src/game/states/state.cpp
@@ -22,13 +22,13 @@ void game_state_machine_destroy(GameStateMachine *sm) {
 
 void game_state_machine_register(GameStateMachine *sm, GameStateID id,
                                   const GameState *state) {
-    if (!sm || id < 0 || id >= GAME_MAX_STATES || !state) return;
+    if (!sm || (int)id < 0 || (int)id >= GAME_MAX_STATES || !state) return;
     sm->states[id] = *state;
 }
 
 void game_state_machine_change(GameStateMachine *sm, GameStateID id,
                                 Agentite_GameContext *ctx) {
-    if (!sm || id < 0 || id >= GAME_MAX_STATES) return;
+    if (!sm || (int)id < 0 || (int)id >= GAME_MAX_STATES) return;
 
     /* Exit current state */
     if (sm->current != GAME_STATE_NONE && sm->states[sm->current].exit) {
